tests/submission/PhilipTaylor/canvas/context.unrecognised.badsuffix.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: context.unrecognised.badsuffix</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.context.html">context</a>.<a href="index.context.unrecognised.html">unrecognised</a>.badsuffix</h1>
<p class="desc">Context name "2d" plus a suffix is unrecognised</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.context.unrecognised">context.unrecognised</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Context name \"2d\" plus a suffix is unrecognised");
_addTest(function(canvas, ctx) {

_assertSame(canvas.getContext("2d#"), null, "canvas.getContext(\"2d#\")", "null");


});
</script>

<div id="log"></div>
</body></html>
